testing out pulling audio from processblock and fixed bug where drawTimeSlices crashed on audio less than 1 second

+++ Source/PluginEditor.h
@@ -69,7 +69,7 @@ private:
     //==============================================================================
     ScopedPointer<Slider> slider;
     ScopedPointer<TextButton> textButton;
-	const Line<float> drawTimeSlice(Rectangle<int> & areaOfOutput, int seconds, int indexInSamples);
+	const Line<float> drawTimeSlice(Rectangle<int> & areaOfOutput, double seconds, int indexInSamples);
 	AudioThumbnail * thumbalina = NULL;
 	const float * arrayOsamps = NULL;
 	File loadFile();
